refactor: 장소 추가 리스트 컴포넌트 분리
